+++ src/model/Producte.java
@@ -1,11 +1,11 @@
 package model;
 
 public abstract  class Producte {
-        protected int preu;
+        protected float preu;
         protected String nom;
         protected String barcode;
 
-    public Producte(int preu, String nom, String barcode) {
+    public Producte(float preu, String nom, String barcode) {
         this.preu = preu;
         this.nom = nom;
         this.barcode = barcode;
@@ -15,7 +15,7 @@ public abstract  class Producte {
 
     // Getters
 
-    public int getPreu() {
+    public float getPreu() {
         return preu;
     }
